Reduce the fetch-toggles-interval to 1s in ephemeral
In the EE we need to reduce the amount of time the Unleash client
takes to refresh the toggle state within the service. This will
need to be done for any service that uses toggles in IQE in order
to avoid using sleeps to compensate (default 10 seconds for a toggle
refresh to occur for a quarkus app).