SWATCH-2820: Run DB migrations independently
This patch creates a new clowdapp, swatch-database, that is now the
controlling entity for all database migrations.  Migrations from the
main project and swatch-contracts have been consolidated under the
swatch-database project.

The migrations are run on OpenShift via an extremely simple Quarkus
application that pulls in the quarkus-liquibase module and defines two
data sources: core and contracts.  The application starts, runs the
migrations, then quits.  In Openshift, we assign this container to a job
and then for each ClowdApp that's dependent on the database, we add an
initContainer that uses the "k8s-wait-for" container to block until the
migration job completes successfully.

The swatch-database ClowdApp now controls the Clowder provisioned
database so all the other ClowdApps needed to be updated to depend on
swatch-database.  The new database is named "swatch-database-db" rather
than "swatch-tally-db".

The gradle Liquibase plugin functionality still works as well.

Unfortunately, changing the relative path of the migration files causes
some issues.  The relative path is part of the changeset record.  So
changing the path make Liquibase think the migration has never run.
Potentially, we could modify the changelog table or set MARK_RAN on the
changesets that have already run, but ultimately I decided that involves
some risk.  The status quo is that contracts changesets live
under the "db" directory and core changesets live under "liquibase".
This state of affairs is unattractive but innocuous.