SWATCH-2820: Run DB migrations independently
This patch creates a new clowdapp, swatch-database, that is now the
controlling entity for all database migrations.  Migrations from the
main project and swatch-contracts have been consolidated under the
swatch-database-migrations project.  The migrations are kept in their
own project to allow other projects to pull them in as test dependencies
and prevent classpath pollution from the MigrationService.

The migrations are run on OpenShift via an extremely simple Java
application (MigrationService) that wraps the Liquibase command line.

In Openshift, we assign this application's container to a job and then
for each ClowdApp that's dependent on the database, we add an
initContainer that uses the "k8s-wait-for" container to block until the
migration job completes successfully.  Each job is given a unique suffix
since if a ClowdJobInvocation with the same name is applied again, the
job doesn't trigger.  The generated suffix is applied as a label to the
ClowdJobInvocation and exposed in a ConfigMap so that other ClowdApps
can reference it.

The swatch-database ClowdApp now controls the Clowder provisioned
database so all the other ClowdApps needed to be updated to depend on
swatch-database.  The new database is named "swatch-database-db" rather
than "swatch-tally-db".

Unfortunately, changing the relative path of the migration files causes
some issues.  The relative path is part of the changeset record.  So
changing the path make Liquibase think the migration has never run.
Potentially, we could modify the changelog table or set MARK_RAN on the
changesets that have already run, but ultimately I decided that involves
some risk.  The status quo is that contracts changesets live
under the "db" directory and core changesets live under "liquibase".
This state of affairs is unattractive but innocuous.

Liquibase Plugin
----------------
The Liquibase Gradle plugin is removed in this commit

The functionality of the plugin is duplicated by the MigrationService
class.  The plugin is awkward to use since passing in Liquibase
arguments requires the syntax
`-PliquibaseExtraArguments='--some=thing'`.

Now Liquibase can just be invoked directly, like so:

```
./gradlew :swatch-database:run --args "status --log-level=WARN"
```

If the first value in `--args` matches the name of one of our migration
contexts (represented in the MigrationService.Context enum), the command
will only apply to that context.  Otherwise, the command will be applied
to all contexts.  For example,

```
./gradlew :swatch-database:run --args "core update --log-level=WARN"
```

Specifying Runtime Properties
-----------------------------
This patch uses standard DATABASE_* arguments instead of
LIQUIBASE_COMMAND_*

It pushes those arguments into system properties so that developers can
still easily control Liquibase through command line arguments.  The
hierarchy of options is as follows (from lowest to highest).

* A properties file with useful defaults for most developer use cases
* DATABASE_* environment variables that will normally be used during
  ClowdApp deployments
* Any option passed via the command line which will give developers
  fine-grained control.

See https://docs.liquibase.com/parameters/working-with-command-parameters.html