SWATCH-2820: Run DB migrations independently
This patch creates a new clowdapp, swatch-database, that is now the
controlling entity for all database migrations.  Migrations from the
main project and swatch-contracts have been consolidated under the
swatch-database-migrations project.  The migrations are kept in their
own project to allow other projects to pull them in as test dependencies
and prevent classpath pollution from the MigrationService.

The migrations are run on OpenShift via an extremely simple Java
application (MigrationService) that wraps the Liquibase command line.

In Openshift, we assign this application's container to be the init
container for any ClowdApp that will need the database.  As the pods
come up, the first init container for a given set of migrations will
grab the changelog lock and run the migrations.  Other pods relying on
the same set of migrations will then have their init containers start,
grab the lock, do nothing (since the migrations have already been run),
and release the lock.

Eventually, all the pods will come up although there will be a bit of
crash-looping when init containers try to run migrations before the
database is available.

The swatch-database ClowdApp now controls the Clowder provisioned
database so all the other ClowdApps needed to be updated to depend on
swatch-database.  The new database is named "swatch-database-db" rather
than "swatch-tally-db".

Unfortunately, changing the relative path of the migration files causes
some issues.  The relative path is part of the changeset record.  So
changing the path make Liquibase think the migration has never run.
Potentially, we could modify the changelog table or set MARK_RAN on the
changesets that have already run, but ultimately I decided that involves
some risk.  The status quo is that contracts changesets live
under the "db" directory and core changesets live under "liquibase".
This state of affairs is unattractive but innocuous.

Liquibase Plugin
----------------
The Liquibase Gradle plugin is removed in this commit

The functionality of the plugin is duplicated by the MigrationService
class.  The plugin is awkward to use since passing in Liquibase
arguments requires the syntax
`-PliquibaseExtraArguments='--some=thing'`.

Now Liquibase can just be invoked directly, like so:

```
./gradlew :swatch-database:run --args "status --log-level=WARN"
```

If the first value in `--args` matches the name of one of our migration
contexts (represented in the MigrationService.Context enum), the command
will only apply to that context.  Otherwise, the command will be applied
to all contexts.  For example,

```
./gradlew :swatch-database:run --args "core update --log-level=WARN"
```

Specifying Runtime Properties
-----------------------------
This MigrationService uses our standard DATABASE_* arguments instead of
LIQUIBASE_COMMAND_*

It pushes those arguments into system properties so that developers can
still easily control Liquibase through command line arguments.  The
hierarchy of options is as follows (from lowest to highest).

* A properties file with useful defaults for most developer use cases
* DATABASE_* environment variables that will normally be used during
  ClowdApp deployments
* Any option passed via the command line which will give developers
  fine-grained control.

See
https://docs.liquibase.com/parameters/working-with-command-parameters.html