Remove Liquibase Gradle plugin
The functionality of this plugin is duplicated by the MigrationService
class.  The plugin is awkward to use since passing in Liquibase
arguments requires the syntax
`-PliquibaseExtraArguments='--some=thing'`.

Now Liquibase can just be invoked directly, like so:

```
./gradlew :swatch-database:run --args "status --log-level=WARN"
```

If the first value in `--args` matches the name of one of our migration
contexts (represented in the MigrationService.Context enum), the command
will only apply to that context.  Otherwise, the command will be applied
to all contexts.  For example,

```
./gradlew :swatch-database:run --args "core update --log-level=WARN"
```