SWATCH-3370: Host tally buckets are not updating the measurement type
We were not updating the measurement type (the physical or virtual...)
and because the measurement type is not part of the bucket key, this was not removed.

After adding b.setMeasurementType(bucket.getMeasurementType());  and syncing the org again, I see this host as physical as expected according to the HBI facts.