Reduce layer churn during builds
By excluding some of the directories that are constantly being updated
due to IDE activity or Gradle activity, we can take better advantage of
layer caching in our builds.

With this change in place, the `COPY . .` command in most of our
Dockerfiles no longer has to rerun almost every time.  On my system,
this brings the total runtime of `bin/build-images.sh` from 9m30s down
to 8m40s.  It's a modest savings, but a welcome one.