SWATCH-3364: Improve troubleshooting when hosts didn't match any products
In order to improve the troubleshooting and to avoid duplicating logic here and there, I had to refactor the ProductNormalizer class to split out the rules into ProductRule implementations.

This way I can loop over these rules to:
- aggregate the matching product tags as done before these changes.
- aggregate the configured product tags to detect when a product tag was a candidate for the host or not

Also, these changes improve the responsibility segregation chain of the product normalizer. Before, this normalizer contained different logic/rules to aggregate the products, and to reconcile the matching products. Now, it delegates the aggregation to the product rules as mentioned before.