Break swatch-database into two parts: application and data
Having the application code mixed together with the database migrations
in swatch-database made it difficult to prevent classpath pollution.
The latest problem that initiated this patch was the inclusion of the
application portion's logback.xml file into the test classpath for other
projects.  This resulted in the test console log being spammed since the
logging settings for the application are much more liberal than those
meant to be used during testing.