Append code coverage report to the test result CI comment
As a time saving measure, we don't run code coverage in CI unless all
the unit tests pass. If we don't run coverage, the link we build in the
comment to the coverage report will be 404. Right now, the unit test
results and the code coverage results are added to the PR in two
different sticky comments to work around this limitation. This is okay,
but it would be nicer just to have a single comment with everything in
it and no 404ed links. So this commit fixes it by having the
sticky-comment action append the coverage report when it succeeds.

You'll notice there's an horizontal rule between the test results and
the coverage report. That's because the last part of the test results is
HTML and without the --- and the very end, the appended message is
interpreted as HTML as well.