Add "topk" to rosa template for ACM dedup
When there are multiple results for metadataMetric to join on, pick the first one's labels to use.

ACM hub/worker node instance ids return multiple records for an instance id with a slight variation in the `service` label, resulting in the inability to join the metric result with the metadata label results.