SWATCH-3530: Use branch to resolve product configuration
Jenkins PR provides the following "GIT_COMMIT" property to get the full sha commit, however it does not work when using it with the GitHub API because this commit sha belongs to a local cloned repository.

Moreover, we can't use the git log or any other git command utility for the same reason than above.

Therefore, the only option is to use the branch name which is provided by Jenkins by GIT_BRANCH.