SWATCH-2579: Filter Capacity API by the most recent active subscription
These changes affect to the following endpoints:

- /api/rhsm-subscriptions/v1/subscriptions/products/{product}
- /api/rhsm-subscriptions/v2/subscriptions/products/{product}
- /api/rhsm-subscriptions/v1/capacity/products/{product}/{metric}

in swatch-api and swatch-contracts.

The root cause of the problem is that we might have multiple active subscriptions, and we're double counting them even when the subscription number (subscription ID) is the same.

The general fix is to always select the most recent active subscription (the one with earliest start_date).

Note that to get the actual subscriptions values (API: /api/rhsm-subscriptions/v2/subscriptions/products/), this was already working fine, but I found that we were not filtering by start_date when getting the measurements. Though this is not affecting the data we return, we save a lot of data when querying subscriptions.

About the rest of the endpoints, I added unit tests to validate the new behaviour plus manual verification.