SWATCH-3527: Convert DebugClientLogger to use standard the JAX-RS API
Replaced Quarkus-based ClientLogger API with standard JAX-RS API: ClientRequestFilter, ClientResponseFilter, which works when using the resteasy rest-client framework.

Note that I added a new property "rest-client-debug-logging.log-responses" to log also the responses, since this is a very dangerous operation (we now need to copy the response in a temporal stream, log it, and copy it back to the original stream, so it's consumed by the following handlers).