Remove kubernetes repo from swatch-database
In an earlier incarnation, swatch-database used a script to monitor the
progress of a Job.  The script called kubectl, so the container included
the Kubernetes RPM repository.  When I removed the job-polling
functionality, I neglected to remove the inclusion of the third-party
repository.  Including this repository causes failures in our build
system since the repo is not approved.

This patch removes the traces of the job-polling functionality that
manifested in the Dockerfile.