Adjust database fallback name
Clowder allows a fallback name for the database.  With the deployment of
swatch-database we want to fall back to our old database name initially
and then change the name in RDS in the future.  This process will allow
us to avoid any outage.  I was not sure if the fallback name would be
"swatch-tally-db" or "swatch-tally".  The former did not work, so this
commit tries the latter.