SWATCH-3663: Disable all DEV services
The Quarkus services are randomly failing when starting the Splunk dev service which we have it disabled using the following property:
```
quarkus.log.handler.splunk.devservices.enabled=false
```

Yet, sometimes the splunk container was trying to start.

Checking the source code of the splunk Quarkus extension, this property is checked when the dev service is executed which might cause inconsistencies.

On the other hand, if we disable all the DEV services using `quarkus.devservices.enabled`, the Splunk Dev Service should not be executed at all.