Enable/disable RBAC role augmentation
Added an ENFORCE_RBAC_ROLES env variable that will
enforce whether RBAC roles will be verified (true)
or not. When false, all roles will be given to the
principal.

The RBAC_ENABLED env var simply enables the augmenter,
allowing the library to bypass RBAC lookups when an
application does not have it enabled. When set to false,
the identity is just passed through without augmenting
roles.