Add Liquibase preconditions to tables controlled by swatch-contracts
PR #4919 introduced a bug where Liquibase would not run correctly in a
swatch database where the changesets had previously been executed. The
run failed with an error that the offering table already exists. I
thought I had addressed this issue by making sure the "contracts"
changesets would run before the "tally" changesets. I didn't apply a
precondition to the legacy changesets because I thought it would alter
the checksums, and dealing with altered checksums in Liquibase can be a
headache. After the bug report, I realized the precondition is necessary
and much to my surprise, they can be added without altering the
checksum.