Validate the event produced by the template so it fails early
Will throw an exception if the data provided by the template is invalid.
This is primarily triggered by the normalization of the facts as they are
defined by HBI as a Map<String, Object> which isn't ideal.

This ensures the test will fail early and before sending the event message.