Fix test-all-services script to work with swatch-api
swatch-api is a bit of a special case.  It's a pseudo-service in that
it's swatch-tally running under a different profile.  The existing
test-all-services script did not handle swatch-api so I added swatch-api
as a Makefile target so it's a first-order service now.

I also deleted a bunch of trailing whitespace and modified the service
test scripts to run under "strict" mode which is how the rest of our
shell scripts operate.