SWATCH-3978: Fix CreateUpdateEventIngestionTest test which was always failing the first time
(1) kafkaBridge.waitForKafkaMessage(Topics.HBI_EVENT_IN, MessageValidators.hbiEventEquals(hbiEvent), 1); -> it does NOT guarantee that the service has consumed the message.
(2) the service will only consume the latest messages. This means that if the message is sent before the service consumer in kafka is reconciled, it will never be consumed. Fixed by adding:

```
%dev.mp.messaging.incoming.hbi-events-in.auto.offset.reset=earliest
```