chore: use grateful destroy of processes in Quarkus local mode
When starting/stopping the services when running the component tests, we were using the destroyForcibly method to immediatelly kill the background process. This was causing that some tests have to wait up to 45 seconds to start consuming new Kafka messages.

The reason is that when a Quarkus service terminates, it sends a signal to the Kafka broker to leave the consumer group. But since we're forcibly killing the process, Quarkus was not sending this signal. Then, when starting the new Quarkus service, Kafka thinks that there is already a consumer and it does not realise that this consumer is gone until 45 seconds (default timeout).

If we grateful shutdown the process, then everything works like expected.

We can test these changes using the following branches: https://github.com/RedHatInsights/rhsm-subscriptions/pull/5114 (running all the swatch contract component tests) or https://github.com/RedHatInsights/rhsm-subscriptions/pull/5135 (running all the swatch metrics hbi component tests).