Fix flush request lock issue
When a synchronous request is made but synchronous operations are disabled,
the lock was acquired and an exception is thrown. When this happened,
the lock was never released since it never hits the flush() call.

This patch ensures that the lock is only aquired when the flush operation
is actually allowed, and will release it when an exception occurs outside
of the flush.