SWATCH-4288: Remove netty-codec-http2 dependency from pom.xml
This dependency was added a while ago to prevent one CVE issue, but this issue was already fixed for netty 4.1.

And having netty 4.2.6.Final forces Spring Boot/Quarkus to use this dependency instead of the ones that these frameworks support it.

For example, Quarkus does not support 4.2.6.Final, but 4.1.128.Final which is causing the traces to fail when being exported.